requirements.txt
pyinstaller
discord.py
fuzzywuzzy
